+++ src/ui/about.rs
@@ -70,16 +70,19 @@ impl SimpleComponent for AboutDialog {
                 "Deutsch — @caem",
                 "Français — @zeGolem https://github.com/zeGolem",
                 "Español — Lautaro Garavano https://github.com/Rattlehead15",
+                "Português — @kafushy",
+                "Italiano — @QuazarOmega https://github.com/quazar-omega",
                 "Türkçe — @Kaozix https://github.com/Kaozix1776",
                 "Türkçe — Kayra Nachfolger https://github.com/kayranachfolger",
-                "Italiano — @QuazarOmega https://github.com/quazar-omega",
-                "Indonesia — @yumekarisu https://github.com/yumekarisu",
-                "简体中文 — Caibin Chen https://github.com/tigersoldier",
-                "日本語 — @zozonteq https://github.com/zozonteq",
+                "Polski — Dominik Opyd https://github.com/oritwoen",
                 // Hungarian?
                 "Svenska — Jakob Fridesjö https://github.com/jakobfridesjo",
-                "Português — @kafushy",
-                "Polski — Dominik Opyd https://github.com/oritwoen"
+                "Nederlands — @everlyy https://github.com/everlyy",
+                "简体中文 — Caibin Chen https://github.com/tigersoldier",
+                "日本語 — @zozonteq https://github.com/zozonteq",
+                "한국어 — @project-dy https://github.com/project-dy",
+                "Indonesia — @yumekarisu https://github.com/yumekarisu",
+                "Tiếng Việt — Nguyễn Hữu Chánh https://github.com/Chanhnh"
             ].join("\n"),
 
             set_debug_info: &[
@@ -97,28 +100,20 @@ impl SimpleComponent for AboutDialog {
                 "<p>Added</p>",
 
                 "<ul>",
-                    "<li>Added feature to map wine drives</li>",
-                    "<li>Added `%launch_args%` magic word for game launching command</li>",
-                    "<li>Added `--session <name>` flag to switch active session</li>",
-                    "<li>Added Portuguese</li>",
-                    "<li>Added Polish</li>",
-                "</ul>",
-
-                "<p>Fixed</p>",
-
-                "<ul>",
-                    "<li>Fixed logo size in the first run window</li>",
+                    "<li>Added Vietnamese</li>",
+                    "<li>Added Korean</li>",
+                    "<li>Added Dutch</li>",
+                    "<li>Made free space checks resolve symlinks</li>",
+                    "<li>Added voiceovers support</li>",
+                    "<li>Added new *Concerning* patch status</li>",
                 "</ul>",
 
                 "<p>Changed</p>",
 
                 "<ul>",
-                    "<li>Updated Turkish</li>",
-                    "<li>Updated Italian</li>",
+                    "<li>Updated development libraries versions</li>",
                     "<li>Updated Japanese</li>",
-                    "<li>Updated Swedish</li>",
-                    "<li>Improved files migration code</li>",
-                    "<li>Updated wishes url extractor to use new cache storage</li>",
+                    "<li>Updated Chinese</li>",
                 "</ul>",
             ].join("\n"),
 
@@ -131,7 +126,7 @@ impl SimpleComponent for AboutDialog {
             connect_close_request[sender] => move |_| {
                 sender.input(AboutDialogMsg::Hide);
 
-                gtk::Inhibit(false)
+                gtk::glib::Propagation::Proceed
             }
         }
     }