+++ src/ui/preferences/general/mod.rs
@@ -1,5 +1,10 @@
 use relm4::prelude::*;
-use relm4::component::*;
+
+use relm4::factory::{
+    AsyncFactoryVecDeque,
+    AsyncFactoryComponent,
+    AsyncFactorySender
+};
 
 use gtk::prelude::*;
 use adw::prelude::*;
@@ -23,7 +28,81 @@ use crate::ui::preferences::main::PreferencesAppMsg;
 use crate::i18n::*;
 use crate::*;
 
+#[derive(Debug)]
+struct VoicePackageComponent {
+    locale: VoiceLocale,
+    installed: bool,
+    sensitive: bool
+}
+
+#[relm4::factory(async)]
+impl AsyncFactoryComponent for VoicePackageComponent {
+    type Init = (VoiceLocale, bool);
+    type Input = GeneralAppMsg;
+    type Output = GeneralAppMsg;
+    type CommandOutput = ();
+    type ParentWidget = adw::ExpanderRow;
+
+    view! {
+        root = adw::ActionRow {
+            set_title: &tr!(&self.locale.to_name().to_ascii_lowercase()),
+
+            add_suffix = &gtk::Button {
+                #[watch]
+                set_visible: self.installed,
+
+                #[watch]
+                set_sensitive: self.sensitive,
+
+                set_icon_name: "user-trash-symbolic",
+                add_css_class: "flat",
+                set_valign: gtk::Align::Center,
+
+                connect_clicked[sender, index] => move |_| {
+                    sender.input(GeneralAppMsg::RemoveVoicePackage(index.clone()));
+                }
+            },
+
+            add_suffix = &gtk::Button {
+                #[watch]
+                set_visible: !self.installed,
+
+                #[watch]
+                set_sensitive: self.sensitive,
+
+                set_icon_name: "document-save-symbolic",
+                add_css_class: "flat",
+                set_valign: gtk::Align::Center,
+
+                connect_clicked[sender, index] => move |_| {
+                    sender.input(GeneralAppMsg::AddVoicePackage(index.clone()));
+                }
+            }
+        }
+    }
+
+    async fn init_model(
+        init: Self::Init,
+        _index: &DynamicIndex,
+        _sender: AsyncFactorySender<Self>,
+    ) -> Self {
+        Self {
+            locale: init.0,
+            installed: init.1,
+            sensitive: true
+        }
+    }
+
+    async fn update(&mut self, msg: Self::Input, sender: AsyncFactorySender<Self>) {
+        self.installed = !self.installed;
+
+        sender.output(msg)
+            .unwrap();
+    }
+}
+
 pub struct GeneralApp {
+    voice_packages: AsyncFactoryVecDeque<VoicePackageComponent>,
     migrate_installation: Controller<MigrateInstallationApp>,
     components_page: AsyncController<ComponentsPage>,
 
@@ -43,6 +122,13 @@ pub enum GeneralAppMsg {
     /// Supposed to be called automatically on app's run when the latest UnityPlayer patch version
     /// was retrieved from remote repos
     SetMainPatch(Option<(Version, JadeitePatchStatusVariant)>),
+
+    // If one ever wish to change it to accept VoiceLocale
+    // I'd recommend to use clone!(@strong self.locale as locale => move |_| { .. })
+    // in the VoicePackage component
+    AddVoicePackage(DynamicIndex),
+    RemoveVoicePackage(DynamicIndex),
+    SetVoicePackageSensitivity(DynamicIndex, bool),
 
     UpdateDownloadedWine,
     UpdateDownloadedDxvk,
@@ -217,6 +303,12 @@ impl SimpleAsyncComponent for GeneralApp {
                     }
                 },
 
+                #[local_ref]
+                voice_packages -> adw::ExpanderRow {
+                    set_title: &tr!("game-voiceovers"),
+                    set_subtitle: &tr!("game-voiceovers-description")
+                },
+
                 gtk::Box {
                     set_orientation: gtk::Orientation::Horizontal,
                     set_spacing: 8,
@@ -309,7 +401,8 @@ impl SimpleAsyncComponent for GeneralApp {
                                 JadeitePatchStatusVariant::Verified => &["success"],
                                 JadeitePatchStatusVariant::Unverified => &["warning"],
                                 JadeitePatchStatusVariant::Broken => &["error"],
-                                JadeitePatchStatusVariant::Unsafe => &["error"]
+                                JadeitePatchStatusVariant::Unsafe => &["error"],
+                                JadeitePatchStatusVariant::Concerning => &["error"]
                             }
 
                             None => &[]
@@ -321,6 +414,9 @@ impl SimpleAsyncComponent for GeneralApp {
                                 JadeitePatchStatusVariant::Unverified => tr!("patch-testing-tooltip"),
                                 JadeitePatchStatusVariant::Broken => tr!("patch-broken-tooltip"),
                                 JadeitePatchStatusVariant::Unsafe => tr!("patch-unsafe-tooltip"),
+
+                                // TODO: special tooltip for concerning status
+                                JadeitePatchStatusVariant::Concerning => tr!("patch-unsafe-tooltip"),
 
                                 _ => String::new()
                             }
@@ -451,7 +547,11 @@ impl SimpleAsyncComponent for GeneralApp {
     ) -> AsyncComponentParts<Self> {
         tracing::info!("Initializing general settings");
 
-        let model = Self {
+        let mut model = Self {
+            voice_packages: AsyncFactoryVecDeque::builder()
+                .launch_default()
+                .forward(sender.input_sender(), std::convert::identity),
+
             migrate_installation: MigrateInstallationApp::builder()
                 .launch(())
                 .detach(),
@@ -467,6 +567,14 @@ impl SimpleAsyncComponent for GeneralApp {
             languages: SUPPORTED_LANGUAGES.iter().map(|lang| tr!(format_lang(lang).as_str())).collect()
         };
 
+        for package in VoiceLocale::list() {
+            model.voice_packages.guard().push_back((
+                *package,
+                CONFIG.game.voices.iter().any(|voice| VoiceLocale::from_str(voice) == Some(*package))
+            ));
+        }
+
+        let voice_packages = model.voice_packages.widget();
         let components_page = model.components_page.widget();
 
         let widgets = view_output!();
@@ -486,6 +594,63 @@ impl SimpleAsyncComponent for GeneralApp {
                 self.main_patch = patch;
             }
 
+            #[allow(unused_must_use)]
+            GeneralAppMsg::AddVoicePackage(index) => {
+                if let Some(package) = self.voice_packages.get(index.current_index()) {
+                    if let Ok(mut config) = Config::get() {
+                        if !config.game.voices.iter().any(|voice| VoiceLocale::from_str(voice) == Some(package.locale)) {
+                            config.game.voices.push(package.locale.to_code().to_string());
+
+                            Config::update(config);
+    
+                            sender.output(PreferencesAppMsg::UpdateLauncherState);
+                        }
+                    }
+                }
+            }
+
+            #[allow(unused_must_use)]
+            GeneralAppMsg::RemoveVoicePackage(index) => {
+                if let Some(package) = self.voice_packages.guard().get_mut(index.current_index()) {
+                    if let Ok(mut config) = Config::get() {
+                        package.sensitive = false;
+
+                        config.game.voices.retain(|voice| VoiceLocale::from_str(voice) != Some(package.locale));
+
+                        Config::update(config.clone());
+
+                        let package = VoicePackage::with_locale(package.locale, config.launcher.edition).unwrap();
+                        let game_path = config.game.path.for_edition(config.launcher.edition).to_path_buf();
+
+                        if package.is_installed_in(&game_path) {
+                            std::thread::spawn(move || {
+                                if let Err(err) = package.delete_in(game_path) {
+                                    tracing::error!("Failed to delete voice package: {:?}", package.locale());
+
+                                    sender.input(GeneralAppMsg::Toast {
+                                        title: tr!("voice-package-deletion-error"),
+                                        description: Some(err.to_string())
+                                    });
+                                }
+
+                                sender.input(GeneralAppMsg::SetVoicePackageSensitivity(index, true));
+                                sender.output(PreferencesAppMsg::UpdateLauncherState);
+                            });
+                        }
+
+                        else {
+                            sender.input(GeneralAppMsg::SetVoicePackageSensitivity(index, true));
+                        }
+                    }
+                }
+            }
+
+            GeneralAppMsg::SetVoicePackageSensitivity(index, sensitive) => {
+                if let Some(package) = self.voice_packages.guard().get_mut(index.current_index()) {
+                    package.sensitive = sensitive;
+                }
+            }
+
             GeneralAppMsg::UpdateDownloadedWine => {
                 self.components_page.sender()
                     .send(ComponentsPageMsg::UpdateDownloadedWine)
@@ -503,7 +668,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     self.migrate_installation.widget().set_transient_for(Some(window.widget()));
                 }
 
-                self.migrate_installation.widget().show();
+                self.migrate_installation.widget().present();
             }
 
             GeneralAppMsg::OpenMainPage => unsafe {